requirements.txt
user-agent-parser
currencyconverter
pyspellchecker
beautifulsoup4
googletrans
requests
openai
geopy
flask
sympy
regex